Refactor MapboxMapNode to a separate file.
