+++ extension-compose/src/main/java/com/mapbox/maps/extension/compose/MapboxMap.kt
@@ -10,51 +10,11 @@ import com.mapbox.maps.MapView
 import com.mapbox.maps.MapboxExperimental
 import com.mapbox.maps.MapboxMap
 import com.mapbox.maps.extension.compose.internal.MapApplier
-import com.mapbox.maps.extension.compose.internal.MapNode
 import com.mapbox.maps.extension.compose.internal.MapPreviewPlaceHolder
+import com.mapbox.maps.extension.compose.internal.MapboxMapComposeNode
 import com.mapbox.maps.plugin.gestures.OnMapClickListener
 import com.mapbox.maps.plugin.gestures.OnMapLongClickListener
-import com.mapbox.maps.plugin.gestures.gestures
 import kotlinx.coroutines.awaitCancellation
-
-/**
- * MapboxMapNode to observe/update the input arguments of MapboxMap.
- */
-private class MapboxMapNode(
-  val controller: MapView,
-  var clickListener: OnMapClickListener?,
-  var longClickListener: OnMapLongClickListener?
-) : MapNode {
-  override fun onAttached() {
-    controller.gestures.apply {
-      clickListener?.let {
-        addOnMapClickListener(it)
-      }
-      longClickListener?.let {
-        addOnMapLongClickListener(it)
-      }
-    }
-  }
-
-  override fun onRemoved() {
-    cleanUp()
-  }
-
-  override fun onClear() {
-    cleanUp()
-  }
-
-  private fun cleanUp() {
-    controller.gestures.apply {
-      clickListener?.let {
-        removeOnMapClickListener(it)
-      }
-      longClickListener?.let {
-        removeOnMapLongClickListener(it)
-      }
-    }
-  }
-}
 
 /**
  * Entry point for adding a Mapbox Map instance to the Jetpack Compose UI.
@@ -103,63 +63,33 @@ public fun MapboxMap(
   )
 
   val parentComposition = rememberCompositionContext()
+  val currentMapInitOptions by rememberUpdatedState(mapInitOptions)
+  val currentOnMapClickListener by rememberUpdatedState(onMapClickListener)
+  val currentOnMapLongClickListener by rememberUpdatedState(onMapLongClickListener)
   val currentContent by rememberUpdatedState(content)
 
   LaunchedEffect(Unit) {
-    val composition = Composition(
-      MapApplier(mapView), parentComposition
-    ).apply {
-      setContent {
-        ComposeNode<MapboxMapNode, MapApplier>(
-          factory = {
-            MapboxMapNode(
-              mapView,
-              onMapClickListener,
-              onMapLongClickListener
-            )
-          },
-          update = {
-            // input arguments updater
-            update(onMapClickListener) { listener ->
-              val previousListener = this.clickListener
-              controller.gestures.apply {
-                previousListener?.let {
-                  removeOnMapClickListener(it)
-                }
-                listener?.let {
-                  addOnMapClickListener(it)
-                }
-              }
-              this.clickListener = listener
-            }
-            update(onMapLongClickListener) { listener ->
-              val previousListener = this.longClickListener
-              controller.gestures.apply {
-                previousListener?.let {
-                  removeOnMapLongClickListener(it)
-                }
-                listener?.let {
-                  addOnMapLongClickListener(it)
-                }
-              }
-              this.longClickListener = listener
-            }
-            update(mapInitOptions) {
-              throw IllegalStateException(
-                """
-                Mutating MapInitOptions during composition is not allowed.
-                """.trimIndent()
-              )
-            }
-          }
-        )
-        currentContent?.let { MapboxMapScope.it() }
+    disposingComposition(
+      Composition(
+        MapApplier(mapView), parentComposition
+      ).apply {
+        setContent {
+          MapboxMapComposeNode(
+            currentMapInitOptions,
+            currentOnMapClickListener,
+            currentOnMapLongClickListener
+          )
+          currentContent?.let { MapboxMapScope.it() }
+        }
       }
-    }
-    try {
-      awaitCancellation()
-    } finally {
-      composition.dispose()
-    }
+    )
+  }
+}
+
+private suspend inline fun disposingComposition(composition: Composition) {
+  try {
+    awaitCancellation()
+  } finally {
+    composition.dispose()
   }
 }
